+++ src/features/user-achievements/components/total-card.tsx
