+++ src/features/user-achievements/components/user-mission-achievements.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 import React from "react";
-import { UserMissionAchievements } from "./index";
+import { UserMissionAchievements } from "./user-mission-achievements";
 
 type MissionAchievementSummary = {
   mission_id: string;
